Move pull request api convert to convert package
